Small tweeks
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@1141 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/server/net_preparse.c
@@ -1119,7 +1119,7 @@ void NPP_QWWriteFloat(int dest, float data)	//replacement write func (nq to qw)
 {
 	union {
 		qbyte b[4];
-		short f;
+		float f;
 	} u;
 	u.f = LittleFloat(data);
 	NPP_QWWriteByte(dest, u.b[0]);
